Split route/integration tests out of unit app_test.py
Per the style guide, the TestClient-driven end-to-end tests (routes, auth,
host-lease, paid-list CRUD, bucket endpoints) are integration tests and belong
in a test_*.py file, while app_test.py should hold unit tests of single
functions/methods. Move them into a new integration file
test_remote_service_connector_routes.py and leave the pure-function and
single-layer (ctx, http-ops, supertokens, paid-gate function) unit tests in
app_test.py.

The two cleanup-sweep tests only needed a fake pool backend (they call
run_pool_host_cleanup_sweep directly with an explicit connection), so they now
build it via make_fake_pool_backend() instead of the route test client, keeping
them self-contained unit tests.

Pure code move plus that refactor: the union of collected tests across both
files is identical to before (203 items, verified disjoint).
